+++ packages/server/listenHandler.ts
@@ -1,7 +1,7 @@
 import {us_listen_socket} from 'uWebSockets.js'
 import PROD from './PROD'
 import getGraphQLExecutor from './utils/getGraphQLExecutor'
-import serverHealthChecker from './utils/serverHealthChecker'
+import serverHealthChecker from './utils/ServerHealthChecker2'
 
 const listenHandler = (listenSocket: us_listen_socket) => {
   const PORT = Number(PROD ? process.env.PORT : process.env.SOCKET_PORT)
